- More work on the ChatModerator
- Added `PerformModerationAction` as an internal method in
PluginManager, this will allow us to call this method when a moderation
action is needed to be taken on a user, it will also set the user's last
time they were timed-out. This will be used for the moderation events
and for when a new blacklist is added and the newly blacklisted phrase
has been said in history chat to timeout the users.
- Added the blacklist check function
- Finish one-man-spam moderation filter